# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": { "title": "task_maneger", "version": "0.1.0" },
    "paths": {
        "/categories/": {
            "get": {
                "tags": ["categories"],
                "summary": "Read Categories",
                "operationId": "categories-read_categories",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    },
                                    "type": "array",
                                    "title": "Response Categories-Read Categories"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["categories"],
                "summary": "Create Category",
                "operationId": "categories-create_category",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{category_id}": {
            "get": {
                "tags": ["categories"],
                "summary": "Read Category",
                "operationId": "categories-read_category",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Category Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["categories"],
                "summary": "Update Category",
                "operationId": "categories-update_category",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Category Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["categories"],
                "summary": "Delete Category",
                "operationId": "categories-delete_category",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Category Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/": {
            "get": {
                "tags": ["tasks"],
                "summary": "Read Tasks",
                "operationId": "tasks-read_tasks",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    },
                                    "type": "array",
                                    "title": "Response Tasks-Read Tasks"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["tasks"],
                "summary": "Create Task",
                "operationId": "tasks-create_task",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTask"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{task_id}": {
            "get": {
                "tags": ["tasks"],
                "summary": "Read Task",
                "operationId": "tasks-read_task",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Task Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["tasks"],
                "summary": "Update Task",
                "operationId": "tasks-update_task",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Task Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTask"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["tasks"],
                "summary": "Delete Task",
                "operationId": "tasks-delete_task",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "title": "Task Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "description": { "type": "string", "title": "Description" },
                    "id": { "type": "integer", "title": "Id" }
                },
                "type": "object",
                "required": ["name", "description"],
                "title": "Category"
            },
            "CreateCategory": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "description": { "type": "string", "title": "Description" }
                },
                "type": "object",
                "required": ["name", "description"],
                "title": "CreateCategory"
            },
            "CreateTask": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "description": { "type": "string", "title": "Description" },
                    "category_id": {
                        "type": "integer",
                        "title": "Category Id"
                    },
                    "status": { "type": "string", "title": "Status" }
                },
                "type": "object",
                "required": ["name", "description", "category_id", "status"],
                "title": "CreateTask"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Task": {
                "properties": {
                    "name": { "type": "string", "title": "Name" },
                    "description": { "type": "string", "title": "Description" },
                    "category_id": {
                        "type": "integer",
                        "title": "Category Id"
                    },
                    "status": { "type": "string", "title": "Status" },
                    "id": {
                        "anyOf": [{ "type": "integer" }, { "type": "null" }],
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": ["name", "description", "category_id", "status"],
                "title": "Task"
            },
            "UpdateCategory": {
                "properties": {
                    "name": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Description"
                    }
                },
                "type": "object",
                "title": "UpdateCategory"
            },
            "UpdateTask": {
                "properties": {
                    "name": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Description"
                    },
                    "category_id": {
                        "anyOf": [{ "type": "integer" }, { "type": "null" }],
                        "title": "Category Id"
                    },
                    "status": {
                        "anyOf": [{ "type": "string" }, { "type": "null" }],
                        "title": "Status"
                    }
                },
                "type": "object",
                "title": "UpdateTask"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                { "type": "string" },
                                { "type": "integer" }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}